refactor: code smell 제거, ex4 부모-자식 클래스간 순환 참조 해결
