+++ src/main/java/com/example/refactoring2/ch10/ex4/voyage/RatingUtils.java
@@ -4,6 +4,6 @@ public class RatingUtils {
   private RatingUtils() {}
 
   public static String rating(Voyage voyage, History history) {
-    return Rating.create(voyage, history).value();
+    return RatingFactory.create(voyage, history).value();
   }
 }
