+++ src/main/java/com/example/refactoring2/ch10/ex5/BillingPlan.java
@@ -1,3 +1,3 @@
 package com.example.refactoring2.ch10.ex5;
 
-public class BillingPlan {}
+public record BillingPlan(String planName) {}
