+++ src/main/java/com/example/refactoring2/ch10/ex5/PaymentHistory.java
@@ -1,7 +1,7 @@
 package com.example.refactoring2.ch10.ex5;
 
 public class PaymentHistory {
-  private int weeksDelinquentInLastYear;
+  private final int weeksDelinquentInLastYear;
 
   public PaymentHistory(int weeksDelinquentInLastYear) {
     this.weeksDelinquentInLastYear = weeksDelinquentInLastYear;
