+++ src/test/java/com/example/refactoring2/ch10/ex4/bird/BirdClassifierTest.java
@@ -12,10 +12,10 @@ class BirdClassifierTest {
   void plumagesTest() {
     List<Bird> birds =
         List.of(
-            Bird.createBird(create("이름1", "유럽 제비")),
-            Bird.createBird(create("이름2", "아프리카 제비")),
-            Bird.createBird(create("이름3", "노르웨이 파랑 앵무")),
-            Bird.createBird(create("이름4", "unknown")));
+            BirdFactory.createBird(create("이름1", "유럽 제비")),
+            BirdFactory.createBird(create("이름2", "아프리카 제비")),
+            BirdFactory.createBird(create("이름3", "노르웨이 파랑 앵무")),
+            BirdFactory.createBird(create("이름4", "unknown")));
 
     Map<String, String> actual = BirdClassifier.plumages(birds);
 
@@ -30,10 +30,10 @@ class BirdClassifierTest {
   void speedsTest() {
     List<Bird> birds =
         List.of(
-            Bird.createBird(create("이름1", "유럽 제비")),
-            Bird.createBird(create("이름2", "아프리카 제비")),
-            Bird.createBird(create("이름3", "노르웨이 파랑 앵무")),
-            Bird.createBird(create("이름4", "unknown")));
+            BirdFactory.createBird(create("이름1", "유럽 제비")),
+            BirdFactory.createBird(create("이름2", "아프리카 제비")),
+            BirdFactory.createBird(create("이름3", "노르웨이 파랑 앵무")),
+            BirdFactory.createBird(create("이름4", "unknown")));
 
     Map<String, Integer> actual = BirdClassifier.speeds(birds);
