+++ src/test/java/com/example/refactoring2/ch10/ex7/Example7Test.java
@@ -8,7 +8,7 @@ import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
 import org.junit.jupiter.params.provider.MethodSource;
 
-public class Example7 {
+class Example7Test {
 
   @ParameterizedTest
   @MethodSource("peopleArguments")
